feat: Added new structured modals
